Create game function
This function initializes and plays the game.
When a winning combo is found or nine rounds is
reached, then the game ends - showing the proper
game state message (win or draw).

I also moved every function into the
GameController module. This was to only make a
select amount of functions globally accessible.